the project restructure to MVVM design and rename of file names
