+++ app/src/main/java/com/khandev/qrcodescanner/presentation/activity/IntroductionActivity.kt
@@ -1,4 +1,4 @@
-package com.khandev.qrcodescanner.activity
+package com.khandev.qrcodescanner.presentation.activity
 
 import android.content.Intent
 import android.os.Bundle
