+++ app/src/main/java/com/khandev/qrcodescanner/presentation/adapter/ResultAdapter.kt
@@ -1,4 +1,4 @@
-package com.khandev.qrcodescanner.adapter
+package com.khandev.qrcodescanner.presentation.adapter
 
 import android.content.Context
 import android.content.Intent
@@ -10,8 +10,8 @@ import android.widget.ImageView
 import android.widget.TextView
 import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
-import com.khandev.qrcodescanner.activity.ResultActivity.Companion.copyTextToClipboard
-import com.khandev.qrcodescanner.database.data.QrCodeEntity
+import com.khandev.qrcodescanner.presentation.activity.ResultActivity.Companion.copyTextToClipboard
+import com.khandev.qrcodescanner.data.QrCodeEntity
 import com.khandev.qrcodescanner.R
 
 class ResultAdapter(private val context: Context, private val resultList: List<QrCodeEntity>) :
