+++ app/src/main/res/layout/activity_error.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context="com.khandev.qrcodescanner.activity.ErrorActivity">
+    tools:context="com.khandev.qrcodescanner.presentation.activity.ErrorActivity">
 
     <ImageView
         android:layout_width="wrap_content"
