+++ app/src/main/res/layout/activity_introduction.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context="com.khandev.qrcodescanner.activity.IntroductionActivity">
+    tools:context="com.khandev.qrcodescanner.presentation.activity.IntroductionActivity">
 
 
     <TextView
