+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.khandev.qrcodescanner.activity.MainActivity">
+    tools:context="com.khandev.qrcodescanner.presentation.activity.MainActivity">
 
 
     <TextView
