+++ app/src/main/res/layout/activity_result.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.khandev.qrcodescanner.activity.ResultActivity">
+    tools:context="com.khandev.qrcodescanner.presentation.activity.ResultActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
